Fix bug with user preferences
Currently, user preferences are not created properly on first launch.

This happens because Files.write() is attempting to write a file
"data/preferences.json", but since the directory doesn't exist, it
throws a NoSuchFileException.

We have to inclue in JsonUserPrefsStorage#saveUserPrefs() an additional
call to FileUtil#createIfMissing() in order to create this missing
directory.

User preferences should be written properly to "data/preferences.json"
on first launch (when it doesn't exist), now.